requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pyjwt==2.8.0
python-multipart==0.0.6
pydantic[email]==2.5.3
cloudinary